backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
email-validator==2.1.1
gunicorn==22.0.0
pytest==8.2.0
